WIP - re-arrange helper/util functions
