+++ src/packages/@ncigdc/modern_components/ClinicalAnalysis/ClinicalVariableCard/utils/categorical.js
@@ -1,0 +1,44 @@
+export const makeCategoricalSetFilters = (selectedBuckets, fieldName, filters) => {
+  const bucketFilters = []
+    .concat(selectedBuckets
+      .filter(bucket => bucket.key !== '_missing').length > 0 && [
+      {
+        content: {
+          field: fieldName,
+          value: selectedBuckets
+            .filter(bucket => bucket.key !== '_missing')
+            .reduce((acc, selectedBucket) =>
+              [...acc, ...selectedBucket.keyArray], []),
+        },
+        op: 'in',
+      },
+    ])
+    .concat(selectedBuckets.some(bucket => bucket.key === '_missing') && 
+      [
+        {
+          content: {
+            field: fieldName,
+            value: 'MISSING',
+          },
+          op: 'is',
+        },
+      ]
+    )
+    .filter(item => item);
+
+  return Object.assign(
+    {},
+    filters,
+    bucketFilters.length && {
+      content: filters.content
+        .concat(
+          bucketFilters.length > 1
+            ? {
+              content: bucketFilters,
+              op: 'or',
+            }
+            : bucketFilters[0]
+        ),
+    }
+  );
+};
